+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Keyboard Layout Guide
 - Cron Jobs
 - Location Services: Convenience wrapper for authorization 
+- UserDefaults Property Wrapper
 
 ### Changed
 - Return cached metrics